checkout function admin authorize sendemail

+++ shopping_tutorial/Controllers/CartController.cs
@@ -27,7 +27,7 @@ namespace shopping_tutorial.Controllers
         {
             return View("~/Views/Checkout/Index.cshtml");
         }
-        public async Task<IActionResult> Add(int Id)
+        public async Task<IActionResult> Add(long  Id)
         {
             ProductModel product = await _dataContext.Products.FindAsync(Id);
             List<CartItemModel> cart = HttpContext.Session.GetJson<List<CartItemModel>>("Cart") ?? new List<CartItemModel>();
@@ -45,7 +45,7 @@ namespace shopping_tutorial.Controllers
             TempData["success"] = "Add Item to cart Successfully";
             return Redirect(Request.Headers["Referer"].ToString());
         }
-        public async Task<IActionResult> Decrease(int Id)
+        public async Task<IActionResult> Decrease(long Id)
         {
             List<CartItemModel> cart = HttpContext.Session.GetJson<List<CartItemModel>>("Cart");
             CartItemModel cartItem = cart.Where(c => c.ProductId == Id).FirstOrDefault();
@@ -68,7 +68,7 @@ namespace shopping_tutorial.Controllers
 			TempData["success"] = "Decrease quantity  Item to cart Successfully";
 			return RedirectToAction("Index");
         }
-        public async Task<IActionResult> Increase(int Id)
+        public async Task<IActionResult> Increase(long Id)
         {
             List<CartItemModel> cart = HttpContext.Session.GetJson<List<CartItemModel>>("Cart");
             CartItemModel cartItem = cart.Where(c => c.ProductId == Id).FirstOrDefault();
@@ -91,7 +91,7 @@ namespace shopping_tutorial.Controllers
 			TempData["success"] = "Increase quantity  Item to cart Successfully";
 			return RedirectToAction("Index");
         }
-        public async Task<IActionResult> Remove(int Id)
+        public async Task<IActionResult> Remove(long Id)
         {
             List<CartItemModel> cart = HttpContext.Session.GetJson<List<CartItemModel>>("Cart");
             cart.RemoveAll(p => p.ProductId == Id);
